+++ dev/src/route.js
@@ -64,13 +64,32 @@
                 if (isQuery) {
                     val = values[prop +'_']; //use raw string
                 }
-                isValid = arrayIndexOf(validationRule, val) !== -1;
+                isValid = this._isValidArrayRule(validationRule, val);
             }
             else if (isFunction(validationRule)) {
                 isValid = validationRule(val, request, values);
             }
 
             return isValid; //fail silently if validationRule is from an unsupported type
+        },
+
+        _isValidArrayRule : function (arr, val) {
+            if (typeof val === 'string') {
+                val = val.toLowerCase();
+            }
+
+            var n = arr.length,
+                item,
+                compareVal;
+
+            while (n-- > 0) {
+                item = arr[n];
+                compareVal = (typeof item === 'string')? item.toLowerCase() : item;
+                if (compareVal === val) {
+                    return true;
+                }
+            }
+            return false;
         },
 
         _getParamsObject : function (request) {